Cleanup, rename to Lti camelCase
